requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.1.1
distro==1.9.0
frozenlist==1.4.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
msgpack==1.0.8
multidict==6.0.5
numba==0.60.0
numpy==2.0.2
openai==0.28.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
setuptools==74.0.0
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.5.0.post1
SpeechRecognition==3.10.4
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.6
